requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
flask==3.0.0

# Email functionality
python-dotenv==1.0.1
keyring==24.3.0  # For secure credential storage

# Scheduling
schedule==1.2.0

# Data processing
python-dateutil==2.8.2

# LLM Integration (optional - install only what you need)
# openai==1.12.0  # For OpenAI GPT (uncomment if using)
# google-generativeai==0.3.2  # For Google Gemini (uncomment if using)